refactor(cli): promote registry commands to root level
- ork push    (was ork registry push)
- ork pull    (was ork registry pull)
- ork inspect (was ork registry info)
- ork patterns (was ork registry list)
- ork registry subcommand group removed entirely — no auth commands
  planned; docker login stays as the auth path
- push/pull/inspect/patterns are now self-contained files with their
  own init(); no shared registry.go wiring needed
- push vars renamed: registryPush* → push*
- inspect var renamed: registryInfoMotif → inspectMotif
- 'ork inspect --view' and '--motif' flags carried over unchanged
- 'To pull:' hints updated to ork pull throughout
- shadow flags on subcommands removed — root-level commands inherit
  from rootCmd which already has debug/kubeconfig/katalog/verbose
