+++ cmd/cli/pull_print.go
